back end/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.4
colorama==0.4.6
cryptography==41.0.1
Deprecated==1.2.14
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
Werkzeug==2.3.6
wrapt==1.15.0
